+++ python/pod5/src/pod5/tools/pod5_filter.py
@@ -3,12 +3,10 @@ Tool for subsetting pod5 files into one or more outputs using a list of read ids
 """
 
 from pathlib import Path
-from time import sleep
 from typing import List
 from pod5.tools.polars_utils import PL_DEST_FNAME, PL_READ_ID, PL_UUID_REGEX
 from pod5.tools.utils import (
     DEFAULT_THREADS,
-    PBAR_DEFAULTS,
     collect_inputs,
     init_logging,
     limit_threads,
@@ -17,18 +15,9 @@ from pod5.tools.utils import (
 
 import polars as pl
 
-from tqdm.auto import tqdm
-
-import pod5 as p5
-from pod5.repack import Repacker
-
 from pod5.tools.parsers import prepare_pod5_filter_argparser, run_tool
-from pod5.tools.pod5_subset import (
-    calculate_transfers,
-    parse_sources,
-)
-from pod5.tools.polars_utils import PL_SRC_FNAME
-from pod5.tools.utils import logged
+from pod5.tools.pod5_subset import build_targets_dict
+import lib_pod5 as p5b
 
 
 logger = init_logging()
@@ -65,58 +54,6 @@ def parse_read_id_targets(ids: Path, output: Path) -> pl.LazyFrame:
     return read_ids
 
 
-@logged(log_time=True)
-def filter_reads(dest: Path, sources: pl.DataFrame, duplicate_ok: bool) -> None:
-    """Copy the reads in `sources` into a new pod5 file at `dest`"""
-    repacker = Repacker()
-    with p5.Writer(dest) as writer:
-        output = repacker.add_output(writer, not duplicate_ok)
-
-        # Count the total number of reads expected
-        total_reads = 0
-        for source, reads in sources.group_by(PL_SRC_FNAME):
-            total_reads += len(reads.get_column(PL_READ_ID))
-
-        pbar = tqdm(
-            total=total_reads,
-            unit="Read",
-            desc="Filtering",
-            leave=True,
-            **PBAR_DEFAULTS,
-        )
-
-        active_limit = 5
-
-        # Copy selected reads from one file at a time
-        for source, reads in sources.group_by(PL_SRC_FNAME):
-            assert len(source) == 1
-            src = Path(source[0])
-            read_ids = reads.get_column(PL_READ_ID).unique().to_list()
-            logger.debug(f"Filtering: {src} - n_reads: {len(read_ids)}")
-
-            if len(read_ids) == 0:
-                logger.debug(f"Skipping: {src}")
-                continue
-
-            while repacker.currently_open_file_reader_count >= active_limit:
-                pbar.update(repacker.reads_completed - pbar.n)
-                sleep(0.2)
-
-            with p5.Reader(src) as reader:
-                repacker.add_selected_reads_to_output(output, reader, read_ids)
-                continue
-
-        repacker.set_output_finished(output)
-        while repacker.currently_open_file_reader_count > 0:
-            pbar.update(repacker.reads_completed - pbar.n)
-            sleep(0.1)
-
-        repacker.finish()
-        pbar.close()
-
-    return
-
-
 @logged_all
 def filter_pod5(
     inputs: List[Path],
@@ -148,23 +85,18 @@ def filter_pod5(
     threads = limit_threads(threads)
 
     _inputs = collect_inputs(inputs, recursive, "*.pod5", threads=threads)
-    sources = parse_sources(_inputs, threads=threads)
-    print(f"Found {len(sources.collect())} read_ids from {len(_inputs)} inputs")
 
-    # Map the target outputs to which source read ids they're comprised of
-    transfers = calculate_transfers(
-        sources=sources,
-        targets=targets,
-        missing_ok=missing_ok,
+    targets_dict = build_targets_dict(targets)
+
+    p5b.subset_pod5s_with_mapping(
+        list(_inputs),
+        output,
+        targets_dict,
+        # threads=threads,
+        missing_ok,
+        duplicate_ok,
+        force_overwrite,
     )
-    print(f"Calculated {len(transfers.collect())} transfers")
-
-    # There will only one output from this
-    groupby_dest = transfers.collect().group_by(PL_DEST_FNAME)
-    for dest, sources in groupby_dest:
-        assert len(dest) == 1
-        filter_reads(dest=dest[0], sources=sources, duplicate_ok=duplicate_ok)
-
     return